third_party/WebKit/LayoutTests/fast/forms/autocapitalize.html

<!DOCTYPE html>
<html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#autocapitalization">
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(function() {
    assert_true('autocapitalize' in document.createElement('input'));
}, "Test that the autocapitalize is available on HTMLInputElement.")

test(function() {
    assert_true('autocapitalize' in document.createElement('textarea'));
}, "Test that the autocapitalize is available on HTMLTextAreaElement.")

test(function() {
    assert_true('autocapitalize' in document.createElement('div'));
}, "Test that the autocapitalize is available on div.")

test(function() {
  var elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div') ];

  elements.forEach(function(e) {
    e.autocapitalize = 'on';
    assert_equals(e.autocapitalize, 'sentences');

    e.autocapitalize = 'off';
    assert_equals(e.autocapitalize, 'none');
  });
}, "Test deprecated values of autocapitalize.");

test(function() {
  var elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div') ];
  var knownValues = [ 'none', 'characters', 'words', 'sentences' ];

  elements.forEach(function(e) {
    // Default value.
    assert_equals(e.autocapitalize, '');

    // Empty value.
    e.autocapitalize = '';
    assert_equals(e.autocapitalize, '');
    assert_equals(e.getAttribute('autocapitalize'), '');
    e.setAttribute('autocapitalize', '');
    assert_equals(e.autocapitalize, '');
    assert_equals(e.getAttribute('autocapitalize'), '');
    assert_equals(e.autocapitalize, '');

    // Invalid value.
    e.autocapitalize = 'foo';
    assert_equals(e.autocapitalize, 'sentences');
    assert_equals(e.getAttribute('autocapitalize'), 'foo');
    e.setAttribute('autocapitalize', 'bar');
    assert_equals(e.autocapitalize, 'sentences');
    assert_equals(e.getAttribute('autocapitalize'), 'bar');

    // Default value.
    e.removeAttribute('autocapitalize');
    assert_equals(e.autocapitalize, '');
    assert_equals(e.getAttribute('autocapitalize'), null);

    // Case insensitive.
    e.setAttribute('autocapitalize', 'NoNe');
    assert_equals(e.autocapitalize, 'none');
    assert_equals(e.getAttribute('autocapitalize'), 'NoNe');
    e.autocapitalize = 'WORDS';
    assert_equals(e.autocapitalize, 'words');
    assert_equals(e.getAttribute('autocapitalize'), 'WORDS');

    knownValues.forEach(function(value) {
      e.setAttribute('autocapitalize', value);
      assert_equals(e.autocapitalize, value);
      assert_equals(e.getAttribute('autocapitalize'), value);

      e.removeAttribute('autocapitalize');

      e.autocapitalize = value;
      assert_equals(e.autocapitalize, value);
      assert_equals(e.getAttribute('autocapitalize'), value);

      e.removeAttribute('autocapitalize');
    });
  });
}, "Test reflection of autocapitalize.");

test(function() {
var testData = [ 'text',
                 'search',
                 'email',
                 'url',
                 'tel',
                 'number',
                 'date',
                 'color',
                 'password' ];

  testData.forEach(function(data) {
    const input = document.createElement('input');
    input.type = data;
    assert_equals(input.autocapitalize, '');

    // Verify that wrapping the input element in a form doesn't change the
    // defaults.
    const form = document.createElement('form');
    form.appendChild(input);
    assert_equals(input.autocapitalize, '');
  });
}, "Test that the IDL attribute returns the empty string if the content "
+ "attribute is not set.")

test(function() {
  const testData = [ 'text',
                     'search',
                     'email',
                     'url',
                     'tel',
                     'number',
                     'date',
                     'color',
                     'password' ];

  testData.forEach(function(data) {
    const input = document.createElement('input');
    input.setAttribute('type', data);
    input.setAttribute('autocapitalize', 'sentences');

    assert_equals(input.autocapitalize, 'sentences');
  });
}, "Verify that even input types that are never autocapitalized support the "
+ "IDL interface.")

</script>
</body>
</html>
